move to aws dir
